acc_group, acc_group_member 업로드

+++ src/main/java/com/iiiiii/accountbook/acc_category/command/domain/repository/CommandAccCategoryRepository.java
@@ -3,6 +3,6 @@ package com.iiiiii.accountbook.acc_category.command.domain.repository;
 import com.iiiiii.accountbook.acc_category.command.domain.aggregate.AccCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommandAccCategoryRepository extends JpaRepository<AccCategoryEntity, Long> {
+public interface CommandAccCategoryRepository extends JpaRepository<AccCategoryEntity, Integer> {
 
 }
